+++ src/db/migrations/migrations.js
@@ -5,6 +5,7 @@ import m0000 from './0000_far_thor.sql';
 import m0001 from './0001_absent_azazel.sql';
 import m0002 from './0002_soft_rachel_grey.sql';
 import m0003 from './0003_groovy_wolverine.sql';
+import m0004 from './0004_striped_marvel_apes.sql';
 
   export default {
     journal,
@@ -12,7 +13,8 @@ import m0003 from './0003_groovy_wolverine.sql';
       m0000,
 m0001,
 m0002,
-m0003
+m0003,
+m0004
     }
   }