+++ src/services/sync/index.ts
@@ -1,0 +1,6 @@
+export * from './types';
+export * from './merge';
+export * from './engine';
+export * from './backends/memory';
+export * from './backends/local';
+export * from './backends/rest';